refactor: update imports for reorganized sections
Update imports to use new paths:
- @/next/sections/* for Blocks, Event, Footer, Header
- @/next/components/richtext/* for Component, Serialize
- @/next/components/* for ToastProvider

Remove old component files from previous locations
