+++ src/app/(pages)/layout.tsx
@@ -2,13 +2,14 @@ import { SpeedInsights } from '@vercel/speed-insights/next';
 import { Analytics } from '@vercel/analytics/react';
 
 import { generateMeta } from '@/next/utils/generateMeta';
-import { HeaderComponent } from '@/next/components/header';
-import { FooterComponent } from '@/next/components/footer';
+import { HeaderComponent } from '@/next/sections/Header';
+import { FooterComponent } from '@/next/sections/Footer';
 
 import './globals.css';
 import { AnimationProvider } from '@/next/animation/context';
 import { getSettings } from '@/next/utils/settings';
 import { getFont } from '@/next/utils/fonts';
+import { ToastProvider } from '@/next/components/ToastProvider';
 
 export const metadata = generateMeta();
 
@@ -31,6 +32,7 @@ export default async function PagesLayout({ children }: { children: React.ReactN
         </div>
         <SpeedInsights />
         <Analytics />
+        <ToastProvider />
       </body>
     </html>
   );
